requirements.txt
absl_py==1.4.0+computecanada
appdirs==1.4.4+computecanada
array_record==0.4.0+computecanada
astunparse==1.6.3+computecanada
cachetools==5.3.0+computecanada
certifi==2022.12.7+computecanada
charset_normalizer==3.1.0+computecanada
click==8.1.3+computecanada
cloudpickle==2.2.1+computecanada
contourpy==1.0.7+computecanada
cycler==0.11.0+computecanada
decorator==5.1.1+computecanada
dm_tree==0.1.8+computecanada
docker-pycreds==0.4.0+computecanada
etils==1.4.0+computecanada
fonttools==4.39.3+computecanada
gast==0.4.0+computecanada
gitdb==4.0.10+computecanada
GitPython==3.1.31+computecanada
grpcio==1.51.3+computecanada
h5py==3.8.0+computecanada
idna==3.4+computecanada
imageio==2.31.1+computecanada
importlib_resources==6.0.1+computecanada
jax==0.4.9+computecanada
jaxlib==0.4.7+cuda11.cudnn82.computecanada
joblib==1.2.0+computecanada
Keras-Preprocessing==1.1.2+computecanada
tensorflow==2.14.0
tensorflow-probability==0.22.0
keras==2.14.0
numpy==1.23.5+computecanada
matplotlib
Pillow==9.5.0+computecanada
scipy==1.10.1+computecanada
threadpoolctl==3.1.0+computecanada
typing_extensions==4.5.0+computecanada
urllib3==1.26.15+computecanada
wandb
Werkzeug==2.3.4+computecanada
wrapt==1.14.1+computecanada
zipp==3.16.2+computecanada
